+++ src/library_sdl.js
@@ -2410,6 +2410,7 @@ var LibrarySDL = {
 
   // SDL_Audio
 
+  SDL_OpenAudio__deps: ['$autoResumeAudioContext'],
   SDL_OpenAudio__proxy: 'sync',
   SDL_OpenAudio__sig: 'iii',
   SDL_OpenAudio: function(desired, obtained) {
@@ -2547,6 +2548,7 @@ var LibrarySDL = {
       // since initializing multiple times fails on Chrome saying 'audio resources have been exhausted'.
       SDL.openAudioContext();
       if (!SDL.audioContext) throw 'Web Audio API is not available!';
+      autoResumeAudioContext(SDL.audioContext);
       SDL.audio.nextPlayTime = 0; // Time in seconds when the next audio block is due to start.
 
       // The pushAudio function with a new audio buffer whenever there is new audio data to schedule to be played back on the device.
@@ -2701,10 +2703,12 @@ var LibrarySDL = {
   },
   Mix_Quit: function(){},
 
+  Mix_OpenAudio__deps: ['$autoResumeAudioContext'],
   Mix_OpenAudio__proxy: 'sync',
   Mix_OpenAudio__sig: 'iiiii',
   Mix_OpenAudio: function(frequency, format, channels, chunksize) {
     SDL.openAudioContext();
+    autoResumeAudioContext(SDL.audioContext);
     SDL.allocateChannels(32);
     // Just record the values for a later call to Mix_QuickLoad_RAW
     SDL.mixerFrequency = frequency;