+++ tools/shared.py
@@ -17,6 +17,7 @@ import os
 import re
 import shutil
 import subprocess
+import time
 import sys
 import tempfile
 
@@ -30,7 +31,7 @@ if sys.version_info[0] == 3 and sys.version_info < (3, 5):
 
 from .toolchain_profiler import ToolchainProfiler
 from .tempfiles import try_delete
-from . import jsrun, cache, tempfiles, colored_logger
+from . import cache, tempfiles, colored_logger
 from . import diagnostics
 
 
@@ -214,6 +215,16 @@ def check_call(cmd, *args, **kw):
     exit_with_error("'%s' failed: %s", ' '.join(cmd), str(e))
 
 
+def run_js_tool(filename, jsargs=[], *args, **kw):
+  """Execute a javascript tool.
+
+  This is used by emcc to run parts of the build process that are written
+  implemented in javascript.
+  """
+  command = NODE_JS + [filename] + jsargs
+  return check_call(command, *args, **kw).stdout
+
+
 # Finds the given executable 'program' in PATH. Operates like the Unix tool 'which'.
 def which(program):
   def is_exe(fpath):
@@ -244,6 +255,25 @@ def which(program):
             return exe_file + suffix
 
   return None
+
+
+# Only used by tests and by ctor_evaller.py.   Once fastcomp is removed
+# this can most likely be moved into the tests/jsrun.py.
+def timeout_run(proc, timeout=None, note='unnamed process', full_output=False, note_args=[], throw_on_failure=True):
+  start = time.time()
+  if timeout is not None:
+    while time.time() - start < timeout and proc.poll() is None:
+      time.sleep(0.1)
+    if proc.poll() is None:
+      proc.kill() # XXX bug: killing emscripten.py does not kill it's child process!
+      raise Exception("Timed out: " + note)
+  stdout, stderr = proc.communicate()
+  out = ['' if o is None else o for o in (stdout, stderr)]
+  if throw_on_failure and proc.returncode != 0:
+    raise subprocess.CalledProcessError(proc.returncode, ' '.join(note_args), stdout, stderr)
+  if TRACK_PROCESS_SPAWNS:
+    logging.info('Process ' + str(proc.pid) + ' finished after ' + str(time.time() - start) + ' seconds. Exit code: ' + str(proc.returncode))
+  return '\n'.join(out) if full_output else out[0]
 
 
 def generate_config(path, first_time=False):
@@ -463,17 +493,18 @@ def env_with_node_in_path():
 
 
 def check_node_version():
-  jsrun.check_engine(NODE_JS)
   try:
     actual = run_process(NODE_JS + ['--version'], stdout=PIPE).stdout.strip()
     version = tuple(map(int, actual.replace('v', '').replace('-pre', '').split('.')))
-    if version >= EXPECTED_NODE_VERSION:
-      return True
-    diagnostics.warning('version-check', 'node version appears too old (seeing "%s", expected "%s")', actual, 'v' + ('.'.join(map(str, EXPECTED_NODE_VERSION))))
-    return False
   except Exception as e:
     diagnostics.warning('version-check', 'cannot check node version: %s', e)
     return False
+
+  if version < EXPECTED_NODE_VERSION:
+    diagnostics.warning('version-check', 'node version appears too old (seeing "%s", expected "%s")', actual, 'v' + ('.'.join(map(str, EXPECTED_NODE_VERSION))))
+    return False
+
+  return True
 
 
 def set_version_globals():
@@ -501,8 +532,10 @@ def perform_sanify_checks():
   logger.info('(Emscripten: Running sanity checks)')
 
   with ToolchainProfiler.profile_block('sanity compiler_engine'):
-    if not jsrun.check_engine(NODE_JS):
-      exit_with_error('The configured node executable (%s) does not seem to work, check the paths in %s', NODE_JS, EM_CONFIG)
+    try:
+      run_process(NODE_JS + ['-e', 'console.log("hello")'], stdout=PIPE)
+    except Exception as e:
+      exit_with_error('The configured node executable (%s) does not seem to work, check the paths in %s (%s)', NODE_JS, config_file_location, str(e))
 
   with ToolchainProfiler.profile_block('sanity LLVM'):
     for cmd in [CLANG_CC, LLVM_AR, LLVM_AS, LLVM_NM]:
@@ -843,13 +876,6 @@ def get_cflags(user_args, cxx):
     return c_opts
 
   return c_opts + emsdk_cflags(user_args, cxx)
-
-
-# Utilities
-def run_js(filename, engine=None, *args, **kw):
-  if engine is None:
-    engine = JS_ENGINES[0]
-  return jsrun.run_js(filename, engine, *args, **kw)
 
 
 def expand_byte_size_suffixes(value):
@@ -1540,21 +1566,22 @@ def read_and_preprocess(filename, expand_macros=False):
   #       we only want the actual settings, hence the [1::2] slice operation.
   settings_str = "var " + ";\nvar ".join(Settings.serialize()[1::2])
   settings_file = os.path.join(temp_dir, 'settings.js')
-  open(settings_file, 'w').write(settings_str)
+  with open(settings_file, 'w') as f:
+    f.write(settings_str)
 
   # Run the JS preprocessor
   # N.B. We can't use the default stdout=PIPE here as it only allows 64K of output before it hangs
   # and shell.html is bigger than that!
   # See https://thraxil.org/users/anders/posts/2008/03/13/Subprocess-Hanging-PIPE-is-your-enemy/
-  (path, file) = os.path.split(filename)
-  if not path:
-    path = None
+  dirname, filename = os.path.split(filename)
+  if not dirname:
+    dirname = None
   stdout = os.path.join(temp_dir, 'stdout')
-  args = [settings_file, file]
+  args = [settings_file, filename]
   if expand_macros:
     args += ['--expandMacros']
 
-  run_js(path_from_root('tools/preprocessor.js'), NODE_JS, args, True, stdout=open(stdout, 'w'), cwd=path)
+  run_js_tool(path_from_root('tools/preprocessor.js'), args, True, stdout=open(stdout, 'w'), cwd=dirname)
   out = open(stdout, 'r').read()
 
   return out
@@ -1706,7 +1733,12 @@ else:
     exit_with_error('emscripten config file not found: ' + CONFIG_FILE)
 
 parse_config_file()
-SPIDERMONKEY_ENGINE = fix_js_engine(SPIDERMONKEY_ENGINE, listify(SPIDERMONKEY_ENGINE))
+# Engine tweaks
+if SPIDERMONKEY_ENGINE:
+  new_spidermonkey = SPIDERMONKEY_ENGINE
+  if '-w' not in str(new_spidermonkey):
+    new_spidermonkey += ['-w']
+  SPIDERMONKEY_ENGINE = fix_js_engine(SPIDERMONKEY_ENGINE, new_spidermonkey)
 NODE_JS = fix_js_engine(NODE_JS, listify(NODE_JS))
 V8_ENGINE = fix_js_engine(V8_ENGINE, listify(V8_ENGINE))
 JS_ENGINES = [listify(engine) for engine in JS_ENGINES]
@@ -1744,6 +1776,7 @@ else:
 # 2: Log stdout and stderr of subprocess spawns. Print out subprocess commands that were executed.
 # 3: Log stdout and stderr, and pass VERBOSE=1 to CMake configure steps.
 EM_BUILD_VERBOSE = int(os.getenv('EM_BUILD_VERBOSE', '0'))
+TRACK_PROCESS_SPAWNS = EM_BUILD_VERBOSE >= 3
 
 set_version_globals()
 
@@ -1799,13 +1832,6 @@ ASM_JS_TARGET = 'asmjs-unknown-emscripten'
 WASM_TARGET = 'wasm32-unknown-emscripten'
 
 check_vanilla()
-
-# Engine tweaks
-if SPIDERMONKEY_ENGINE:
-  new_spidermonkey = SPIDERMONKEY_ENGINE
-  if '-w' not in str(new_spidermonkey):
-    new_spidermonkey += ['-w']
-  SPIDERMONKEY_ENGINE = fix_js_engine(SPIDERMONKEY_ENGINE, new_spidermonkey)
 
 Settings = SettingsManager()
 verify_settings()